refactor: change folder structure for Quotas Domain Entities
