+++ Modules/Quotas/src/Quotas.Domain/Aggregates/Entities/Identity.cs
@@ -1,0 +1,5 @@
+﻿namespace Backbone.Modules.Quotas.Domain.Aggregates.Entities;
+
+public class Identity
+{
+}
