+++ Modules/Quotas/src/Quotas.Domain/Aggregates/Entities/IndividualQuota.cs
@@ -1,0 +1,4 @@
+﻿namespace Backbone.Modules.Quotas.Domain.Aggregates.Entities;
+
+public class IndividualQuota : Quota
+{ }
