+++ Modules/Quotas/src/Quotas.Domain/Aggregates/Metrics/Metric.cs
@@ -1,0 +1,8 @@
+﻿using Backbone.Modules.Quotas.Domain.Aggregates.Entities;
+
+namespace Backbone.Modules.Quotas.Domain.Aggregates.Metrics;
+
+public class Metric
+{
+    public MetricStatus MetricStatus { get; set; }
+}
