--- Modules/Quotas/src/Quotas.Domain/Entities/IMetricCalculator.cs
@@ -1,4 +1,0 @@
-﻿namespace Backbone.Modules.Quotas.Domain.Entities;
-public interface IMetricCalculator
-{
-}
