--- Modules/Quotas/src/Quotas.Domain/Entities/Metric.cs
@@ -1,6 +1,0 @@
-﻿namespace Backbone.Modules.Quotas.Domain.Entities;
-
-public class Metric
-{
-    public MetricStatus MetricStatus { get; set; }
-}
