--- Modules/Quotas/src/Quotas.Domain/Entities/Tier.cs
@@ -1,5 +1,0 @@
-﻿namespace Backbone.Modules.Quotas.Domain.Entities;
-
-public class Tier
-{
-}
