--- Modules/Quotas/src/Quotas.Domain/Entities/TierQuota.cs
@@ -1,4 +1,0 @@
-﻿namespace Backbone.Modules.Quotas.Domain.Entities;
-
-public class TierQuota
-{ }
