--- Modules/Quotas/src/Quotas.Domain/Entities/TierQuotaDefinition.cs
@@ -1,4 +1,0 @@
-﻿namespace Backbone.Modules.Quotas.Domain.Entities;
-
-public class TierQuotaDefinition
-{ }
