+++ Modules/Quotas/src/Quotas.Domain/IMetricCalculatorFactory.cs
@@ -1,0 +1,4 @@
+﻿namespace Backbone.Modules.Quotas.Domain;
+public interface IMetricCalculatorFactory
+{
+}
